refactor: rework issue_comment.created event handler
